pyproject.toml
[project]
name = "proj2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.111",
	"uvicorn[standard]>=0.30",
	"pydantic>=2.7",
	"python-multipart>=0.0.9",
	"google-generativeai>=0.7",
	"httpx>=0.27",
	"pandas>=2.2",
	"numpy>=1.26",
	"beautifulsoup4>=4.12",
	"lxml>=5.2",
	"duckdb>=1.0",
	"pyarrow>=16.1",
	"matplotlib>=3.9",
	"pillow>=10.3",
]

[tool.uv]
package = true

[project.scripts]
start = "app.main:dev"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pyright]
typeCheckingMode = "basic"

